+++ Source/EventFlow.TestHelpers/Aggregates/Commands/ThingyThrowExceptionInSagaCommand.cs
@@ -1,4 +1,4 @@
-// The MIT License (MIT)
+﻿// The MIT License (MIT)
 // 
 // Copyright (c) 2015-2020 Rasmus Mikkelsen
 // Copyright (c) 2015-2020 eBay Software Foundation
@@ -21,22 +21,27 @@
 // IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN
 // CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
 
-using System;
+using EventFlow.Commands;
 using System.Threading;
 using System.Threading.Tasks;
 
-namespace EventFlow.Sagas
+namespace EventFlow.TestHelpers.Aggregates.Commands
 {
-    public class SagaErrorHandler : ISagaErrorHandler
+    public class ThingyThrowExceptionInSagaCommand : Command<ThingyAggregate, ThingyId>
     {
-        public Task<bool> HandleAsync(
-            ISagaId sagaId,
-            SagaDetails sagaDetails,
-            Exception exception,
+        public ThingyThrowExceptionInSagaCommand(ThingyId aggregateId) : base(aggregateId)
+        {
+        }
+    }
+
+    public class ThingyThrowExceptionInSagaCommandHandler
+        : CommandHandler<ThingyAggregate, ThingyId, ThingyThrowExceptionInSagaCommand>
+    {
+        public override Task ExecuteAsync(ThingyAggregate aggregate, ThingyThrowExceptionInSagaCommand command,
             CancellationToken cancellationToken)
         {
-            // The default handler cannot handle anything!
-            return Task.FromResult(false);
+            aggregate.RequestSagaException();
+            return Task.FromResult(0);
         }
     }
 }